More implementation of delete_min
Firstly, the old meld function was not really meld. Meld just adds a
tree to the forest. The old meld has been renamed to link and is used in
delete_min.

The print function has proven wrong as it only prints the first tree in
the forest.

The delete_min function is not working yet! Sorry :(